scraper-engine/requirements.txt
CMRESHandler==1.0.0
alembic==1.6.3
blinker==1.4
chemist==1.7.0
click==8.0.1
colorlog==5.0.1
cssselect==1.1.0
defusedxml==0.7.1
elasticsearch==7.13.0
fastapi==0.68.0
lxml==4.6.3
psycopg2==2.8.6
python-dateutil==2.8.1
pyzmq==22.1.0
uiclasses==2.4.0
uvloop==0.16.0
tqdm==4.62.0
python-json-logger==2.0.2
redis==3.5.3
